Add overwrite feature
Add method to check If a file exists. Add checking when the user inputs a name for the spreadsheet file. Add logic to give the user the possibility to overwrite the existing file If there's already a file with that name or create a new one. Add method to create a new name for the spreadsheet file

close #2